+++ src/ModularPipelines/Engine/IModuleResultRepository.cs
@@ -1,11 +1,12 @@
-﻿using ModularPipelines.Models;
+﻿using ModularPipelines.Context;
+using ModularPipelines.Models;
 using ModularPipelines.Modules;
 
 namespace ModularPipelines.Engine;
 
 public interface IModuleResultRepository
 {
-    Task SaveResultAsync<T>(ModuleBase module, ModuleResult<T> moduleResult);
+    Task SaveResultAsync<T>(ModuleBase module, ModuleResult<T> moduleResult, IPipelineContext pipelineContext);
 
-    Task<ModuleResult<T>?> GetResultAsync<T>(ModuleBase module);
+    Task<ModuleResult<T>?> GetResultAsync<T>(ModuleBase module, IPipelineContext pipelineContext);
 }